Move decision_proceduret into solvers/prop
All modules that use decision_proceduret also use solvers.
So, there is no need to have this interface in util.
