+++ src/goto-instrument/accelerate/scratch_program.cpp
@@ -12,7 +12,8 @@ Author: Matt Lewis
 #include "scratch_program.h"
 
 #include <util/fixedbv.h>
-#include <util/decision_procedure.h>
+
+#include <solvers/prop/decision_procedure.h>
 
 #include <goto-symex/slice.h>
