+++ src/solvers/prop/decision_procedure.cpp
