+++ src/solvers/prop/decision_procedure.h
@@ -9,8 +9,8 @@ Author: Daniel Kroening, kroening@kroening.com
 /// \file
 /// Decision Procedure Interface
 
-#ifndef CPROVER_UTIL_DECISION_PROCEDURE_H
-#define CPROVER_UTIL_DECISION_PROCEDURE_H
+#ifndef CPROVER_SOLVERS_PROP_DECISION_PROCEDURE_H
+#define CPROVER_SOLVERS_PROP_DECISION_PROCEDURE_H
 
 #include <iosfwd>
 #include <string>
@@ -62,4 +62,4 @@ inline decision_proceduret &operator<<(
   return dest;
 }
 
-#endif // CPROVER_UTIL_DECISION_PROCEDURE_H
+#endif // CPROVER_SOLVERS_PROP_DECISION_PROCEDURE_H
