+++ src/solvers/prop/prop_conv.h
@@ -13,11 +13,11 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <string>
 #include <map>
 
-#include <util/decision_procedure.h>
 #include <util/expr.h>
 #include <util/message.h>
 #include <util/std_expr.h>
 
+#include "decision_procedure.h"
 #include "literal.h"
 #include "literal_expr.h"
 #include "prop.h"
